project/DjangoDRFTask2/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
flake8==7.1.1
gunicorn==23.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg==3.2.3
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
